fix: Disable lexical tracker for embedded modules
When using `polymorphic_embeds_one` or `polymorphic_embeds_many` we
create compile-time dependencies from the parent schema to the embedded
schemas.

This change expands the embedded schema module names and removes the
lexical tracker to avoid creating this compile-time dependencies. This
is safe to do since in this place we are not calling any functions on
the association modules.

This uses the same approach that was used in Ecto's `embeds_one` and
`embeds_many` associations: elixir-ecto/ecto#1670